Move the Classes from the Internal ProcessHolder to a ClassesHolder which should be available to Plugins.
