+++ AndroidAnnotations/androidannotations-core/androidannotations/src/main/java/org/androidannotations/internal/process/ModelProcessor.java
@@ -27,6 +27,7 @@ import javax.lang.model.element.TypeElement;
 
 import org.androidannotations.handler.AnnotationHandler;
 import org.androidannotations.handler.GeneratingAnnotationHandler;
+import org.androidannotations.helper.ClassesHolder;
 import org.androidannotations.holder.GeneratedClassHolder;
 import org.androidannotations.internal.InternalAndroidAnnotationsEnvironment;
 import org.androidannotations.internal.exception.ProcessingException;
@@ -61,9 +62,11 @@ public class ModelProcessor {
 
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	public ProcessResult process(AnnotationElements validatedModel) throws Exception {
-		ProcessHolder processHolder = new ProcessHolder(environment.getProcessingEnvironment());
+		ProcessHolder processHolder = new ProcessHolder();
+		ClassesHolder classesHolder = new ClassesHolder(environment.getProcessingEnvironment());
 
 		environment.setProcessHolder(processHolder);
+		environment.setClassesHolder(classesHolder);
 
 		LOGGER.info("Processing root elements");
 
@@ -143,7 +146,7 @@ public class ModelProcessor {
 		}
 
 		return new ProcessResult(//
-				processHolder.codeModel(), //
+				classesHolder.codeModel(), //
 				processHolder.getOriginatingElements());
 	}
